+++ src/assembler/include/BinaryGenerationPass.hpp
@@ -7,11 +7,11 @@ class BinaryGenerationPass : public ASMPass{
         BinaryGenerationPass(std::string filename, uint64_t binarySize=0);
         virtual std::vector<Token>& operator() (std::vector<Token>& tokens) override;
 
-        std::vector<int64_t> const& getBinary();
+        std::vector<int8_t> const& getBinary();
 
     private:
         std::string filename;
         
-        std::vector<int64_t> binary;
+        std::vector<int8_t> binary;
         uint64_t binarySize;
 };
